+++ FSR/DigitalTwin/Server/src/FSR.DigitalTwin.Domain/Model/Process/HRC/Task/HRCTask.cs
@@ -7,7 +7,7 @@ public class HRCTask
         Any = 0, Human = 1, Robot = 2
     }
 
-    private readonly long _horizon;
+    private readonly float _horizon;
     public Resource Resource { get; }
     public Resource Type { get; }
     public Resource? Target { set; get; }
@@ -17,15 +17,15 @@ public class HRCTask
     public string? Goal { set; get; }
     public string? Start { set; get; }
     public EAgent Agent { set; get; } = EAgent.Any;
-    public Tuple<long, long> Duration => new(
+    public Tuple<float, float> Duration => new(
         Math.Max(1, AverageDuration - DurationUncertainty),
         Math.Min(AverageDuration + DurationUncertainty, _horizon)
     );
-    public long AverageDuration { set; get; } = 1;
-    public long DurationUncertainty { set; get; }
+    public float AverageDuration { set; get; } = 1;
+    public float DurationUncertainty { set; get; }
     public double SuccessRate { set; get; } = 0.99;
 
-    public HRCTask(Resource function, Resource type, long horizon)
+    public HRCTask(Resource function, Resource type, float horizon)
     {
         Name = function.ToString();
         Description = Name;
